move test for declare kwarg into publisher core
