+++ test/test_messaging.py
@@ -1013,69 +1013,3 @@ class TestPublisherOptionPrecedence(object):
         message = get_message_from_queue(queue.name)
         assert message.properties['expiration'] == str(2 * 1000)
 
-
-# MYB: mostly copied to test/amqp/test_publish::TestPublisher
-@pytest.mark.behavioural
-class TestPublisherOptions(object):
-
-    @pytest.fixture
-    def routing_key(self):
-        return "routing_key"
-
-    @pytest.fixture
-    def exchange(self, amqp_uri):
-        """ Make a "sniffer" queue bound to the exchange receiving messages
-        """
-        exchange = Exchange(name="exchange")
-        with get_connection(amqp_uri) as connection:
-            maybe_declare(exchange, connection)
-        return exchange
-
-    @pytest.fixture
-    def queue(self, amqp_uri, exchange, routing_key):
-        queue = Queue(
-            name="queue", exchange=exchange, routing_key=routing_key
-        )
-
-        with get_connection(amqp_uri) as connection:
-            maybe_declare(queue, connection)
-        return queue
-
-    @pytest.fixture
-    def service_base(self, exchange):
-
-        class Service(object):
-            name = "service"
-
-            publish = Publisher(exchange=exchange)
-
-            @dummy
-            def proxy(self, *args, **kwargs):
-                self.publish(*args, **kwargs)
-
-        return Service
-
-    @patch('kombu.messaging.maybe_declare', wraps=maybe_declare)
-    def test_declare(
-        self, maybe_declare, container_factory, rabbit_config,
-        get_message_from_queue, exchange, queue, service_base, routing_key
-    ):
-        container = container_factory(service_base, rabbit_config)
-        container.start()
-
-        declare = [
-            Queue(name="q1", exchange=exchange, routing_key=routing_key),
-            Queue(name="q2", exchange=exchange, routing_key=routing_key)
-        ]
-
-        with entrypoint_hook(container, "proxy") as publish:
-            publish("payload", routing_key=routing_key, declare=declare)
-
-        assert maybe_declare.call_args_list == [
-            call(exchange, ANY, ANY),
-            call(declare[0], ANY, ANY),
-            call(declare[1], ANY, ANY)
-        ]
-
-        assert get_message_from_queue(declare[0].name).payload == "payload"
-        assert get_message_from_queue(declare[1].name).payload == "payload"
